edit and delete comment tests

+++ app/Livewire/DeleteComment.php
@@ -4,6 +4,7 @@ namespace App\Livewire;
 
 use App\Models\Comment;
 use Livewire\Component;
+use Illuminate\Http\Response;
 
 class DeleteComment extends Component
 {
@@ -19,6 +20,10 @@ class DeleteComment extends Component
 
     public function deleteComment()
     {
+        if (auth()->guest() || auth()->user()->cannot('delete', $this->comment)) {
+            abort(Response::HTTP_FORBIDDEN, 'You are not allowed to update this comment.');
+        }
+
         $this->comment->delete();
         $this->dispatch('comment-was-deleted', "Comment was deleted successfully!");
     }
